requirements.txt
discord.py>=2.0.0
GitPython>=3.1.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
PyGithub>=1.55
pytest>=7.0.0
pytest-asyncio>=0.21.0